# Acceso a Datos/pedidos-mongo-api-server/src/main/java/ceu/dam/ad/mongo/model/PedidoDetalle.java
package ceu.dam.ad.mongo.model;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PedidoDetalle {
	@NotNull
	private Articulo articulo;
	@NotNull
	@Min(1)
	private Integer cantidad;
	
}